Tried to allow changing device to work in cuda graph decoder.
The first part of test_change_devices() works, but unfortunately the
second part does not.

It is possible to change devices if the decoder has not run yet, but
not possible to do so if the decoder has run already.

I'm not sure why. There's an error about doing a synchronous call
right at the line after the comment "# This is failing...", but I have
not managed to debug it.